feat: pass all stores to extenders
