+++ .changeset/healthy-tomatoes-attend.md
@@ -1,0 +1,12 @@
+---
+'@felte/common': minor
+'@felte/core': minor
+'@felte/element': minor
+'felte': minor
+'@felte/preact': minor
+'@felte/react': minor
+'@felte/solid': minor
+'@felte/vanilla': minor
+---
+
+Pass all stores, `createSubmitHandler` and `handleSubmit` to extenders
