+++ packages/site/src/components/Home/HomeExample.astro
@@ -27,9 +27,9 @@
       <span>Email:</span>
       <input name="email" type="email" aria-describedby="signin-description" />
     </label>
-    <span id="signin-description"
-      >Use any value you want, this is just a demo.</span
-    >
+    <span id="signin-description">
+      Use any value you want, this is just a demo.
+    </span>
     <label>
       <span>Password:</span>
       <input name="password" type="password" />
